+++ crates/web-app-dioxus/src/navbar.rs
@@ -9,8 +9,8 @@ use log::warn;
 use valens_domain::{self as domain, SessionService};
 
 use crate::{
-    DOMAIN_SERVICE, DROP_SET_CALCULATOR, ERRORS, METRONOME, NO_CONNECTION, ONE_REP_MAX_CALCULATOR,
-    Route,
+    DOMAIN_SERVICE, DROP_SET_CALCULATOR, METRONOME, NO_CONNECTION, ONE_REP_MAX_CALCULATOR, Route,
+    notification::notify_error,
     ongoing_training_session::OngoingTrainingSession,
     page::common::{
         DropSetCalculator, Metronome, MutableTimer, OneRepMaxCalculator, Stopwatch,
@@ -235,9 +235,7 @@ pub fn Navbar() -> Element {
                                         navigator().push(Route::Login {});
                                     }
                                     Err(err) => {
-                                        ERRORS
-                                            .write()
-                                            .push(format!("Failed to sign out: {err}"));
+                                        notify_error(format!("Failed to sign out: {err}"));
                                         }
                                 }
                                 menu_visible.set(false);