+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Body weight page: Reordered chart legend entries
 - Body fat page: Reordered chart legend entries
 - Navigation bar: Renamed "Log out" to "Sign out"
+- Presentation of error messages from a dialog to a notification below the navigation bar
 
 ### Fixed